+++ packages/labnative/pages/List.jsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import { List } from '@gympass/yoga';
-import { Text } from 'react-native';
+import { Alert } from 'react-native';
+import { Text, List } from '@gympass/yoga';
 
-const data = [
+const listItemData = [
   { key: 'Devin' },
   { key: 'Dan' },
   { key: 'Dominic' },
   { key: 'Jackson' },
   { key: 'James' },
+];
+
+const listSelectableItemData = [
   { key: 'Joel' },
   { key: 'John' },
   { key: 'Jillian' },
@@ -17,14 +20,28 @@ const data = [
 
 const ListPage = () => {
   return (
-    <List
-      data={data}
-      renderItem={({ item }) => (
-        <List.Item>
-          <Text>{item.key}</Text>
-        </List.Item>
-      )}
-    />
+    <>
+      <Text.SectionTitle>List Item</Text.SectionTitle>
+      <List
+        data={listItemData}
+        renderItem={({ item }) => (
+          <List.Item>
+            <Text>{item.key}</Text>
+          </List.Item>
+        )}
+      />
+
+      <Text.SectionTitle>List Selectable Item</Text.SectionTitle>
+      <List
+        scrollEnabled={false}
+        data={listSelectableItemData}
+        renderItem={({ item }) => (
+          <List.Item small onPress={() => Alert.alert('You pressed', item.key)}>
+            <Text>{item.key}</Text>
+          </List.Item>
+        )}
+      />
+    </>
   );
 };