+++ packages/yoga/src/List/native/ListItem.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
+import { TouchableHighlight } from 'react-native';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
+import styled, { withTheme } from 'styled-components';
 
 const StyledView = styled.View(
   ({
     divided,
+    small,
     theme: {
       yoga: {
         components: { list },
@@ -20,26 +22,67 @@ const StyledView = styled.View(
       : ''
   }
 
-    padding:
-      ${list.listItem.padding.top}px
-      ${list.listItem.padding.right}px
-      ${list.listItem.padding.bottom}px
-      ${list.listItem.padding.left}px;
+  ${
+    small
+      ? `
+          padding:
+            ${list.listItem.small.padding.top}px
+            ${list.listItem.small.padding.right}px
+            ${list.listItem.small.padding.bottom}px
+            ${list.listItem.small.padding.left}px;
+        `
+      : `
+          padding:
+            ${list.listItem.padding.top}px
+            ${list.listItem.padding.right}px
+            ${list.listItem.padding.bottom}px
+            ${list.listItem.padding.left}px;
+        `
+  }
   `,
 );
 
-const ListItem = ({ divided, ...rest }) => (
-  <StyledView divided={divided} {...rest} />
-);
+const ListItem = ({
+  theme: {
+    yoga: {
+      components: {
+        list: { listItem },
+      },
+    },
+  },
+  small,
+  divided,
+  onPress,
+  ...rest
+}) => {
+  const Component = <StyledView small={small} divided={divided} {...rest} />;
+
+  if (onPress) {
+    return (
+      <TouchableHighlight
+        onPress={onPress}
+        underlayColor={listItem.selectable.color}
+      >
+        {Component}
+      </TouchableHighlight>
+    );
+  }
+
+  return Component;
+};
 
 ListItem.propTypes = {
+  small: PropTypes.bool,
   divided: PropTypes.bool,
+  onPress: PropTypes.func,
 };
 
 ListItem.defaultProps = {
+  small: false,
   divided: true,
+  onPress: undefined,
 };
 
 ListItem.displayName = 'List.Item';
 
-export default ListItem;
+export default withTheme(ListItem);