+++ packages/yoga/src/Result/native/Attendances.jsx
@@ -6,39 +6,16 @@ import Icon from '../../Icon';
 import Text from '../../Text';
 import Rate from './Rate';
 
-const List = styled(Text.Regular)`
+const List = styled(Text.Tiny).attrs({
+  numberOfLines: 1,
+})`
   flex: 1;
 `;
 
-const ItemSeparator = styled.View`
-  ${({
-    theme: {
-      yoga: {
-        spacing: { zero, xxxsmall },
-      },
-    },
-  }) => {
-    return `
-      width: ${xxxsmall};
-      height: ${zero};
-    `;
-  }}
-`;
-
-const IconWrapper = styled.View`
-  ${({
-    theme: {
-      yoga: {
-        spacing: { small },
-      },
-    },
-  }) => {
-    return `
-      width: ${small};
-      height: ${small};
-    `;
-  }}
-`;
+const ItemSeparator = styled(Box).attrs({
+  width: 'xxxsmall',
+  height: 'zero',
+})``;
 
 const Attendances = ({ attendances, rate }) => (
   <Box
@@ -48,24 +25,17 @@ const Attendances = ({ attendances, rate }) => (
     justifyContent="center"
     flexDirection="row"
   >
-    <List numberOfLines={1} size="xsmall" textAlignVertical="bottom">
-      {attendances &&
-        attendances.map(({ description, icon }) => (
-          <React.Fragment key={description}>
-            <IconWrapper>
-              <Icon
-                as={icon}
-                fill="medium"
-                width="100%"
-                height="100%"
-                style={{ marginTop: 3 }}
-              />
-            </IconWrapper>
-            <ItemSeparator />
-            {description}
-            <ItemSeparator />
-          </React.Fragment>
-        ))}
+    <List>
+      {attendances.map(({ description, icon }) => (
+        <React.Fragment key={description}>
+          <Box mt={-2}>
+            <Icon as={icon} fill="deep" size="xsmall" />
+          </Box>
+          <ItemSeparator />
+          {description}
+          <ItemSeparator />
+        </React.Fragment>
+      ))}
     </List>
     {rate && <Rate rate={rate} />}
   </Box>