+++ packages/yoga/src/Result/native/Details.jsx
@@ -13,54 +13,19 @@ import styled from 'styled-components';
 
 import Text from '../../Text';
 import Icon from '../../Icon';
+import Box from '../../Box';
 
-const Container = styled.Text`
-  text-align-vertical: center;
+const Container = styled(Text.Tiny)`
   flex: 1;
+  height: ${({ theme }) => theme.yoga.spacing.small}px;
 `;
 
-const Wrapper = styled.View`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
+const Separator = styled(Box).attrs({
+  width: 'xxxsmall',
+  height: 'xxxsmall',
+})``;
 
-  ${({
-    theme: {
-      yoga: {
-        spacing: { xxxsmall },
-      },
-    },
-  }) => `
-    margin-top: ${xxxsmall}px;
-  `}
-`;
-const IconWrapper = styled.View`
-  ${({
-    theme: {
-      yoga: {
-        spacing: { small },
-      },
-    },
-  }) => `
-    width: ${small}px;
-    height: ${small}px;
-  `}
-`;
-
-const Separator = styled.View`
-  ${({
-    theme: {
-      yoga: {
-        spacing: { xxxsmall },
-      },
-    },
-  }) => `
-    width: ${xxxsmall}px;
-    height: ${xxxsmall}px;
-  `}
-`;
-
-const StyledText = styled(Text.Small)`
+const StyledText = styled(Text.Tiny)`
   ${({
     theme: {
       yoga: {
@@ -84,7 +49,7 @@ const ResultDetails = ({
   const numberOfItemsLeft = items.length - limit;
 
   return (
-    <Wrapper>
+    <Box flexDirection="row" alignItems="center" mt="xxxsmall">
       <Container numberOfLines={1}>
         {refinedList?.map(
           ({ icon: IconComponent, variant, ...props }, index) => {
@@ -95,15 +60,9 @@ const ResultDetails = ({
               // eslint-disable-next-line react/no-array-index-key
               <React.Fragment key={index}>
                 {IconComponent && (
-                  <IconWrapper>
-                    <Icon
-                      as={IconComponent}
-                      fill={variant}
-                      width={14}
-                      height={14}
-                      style={{ marginTop: 3 }}
-                    />
-                  </IconWrapper>
+                  <Box mr="xxxsmall" mt={-2}>
+                    <Icon as={IconComponent} fill={variant} size="xsmall" />
+                  </Box>
                 )}
                 <Item variant={variant} {...props} />
 
@@ -133,7 +92,7 @@ const ResultDetails = ({
           },
         )}
       </Container>
-    </Wrapper>
+    </Box>
   );
 };
 
@@ -158,7 +117,7 @@ ResultDetails.defaultProps = {
   limit: undefined,
   limitLabel: undefined,
   dots: false,
-  renderItem: Text.Small,
+  renderItem: Text.Tiny,
 };
 
 export default ResultDetails;